Login Page Website
To create the website, I started by designing the layout using HTML and CSS. I used HTML to create the basic structure of the website, including the header, footer, and main content section. 

I then used CSS to style the website, including the colors, fonts, and layout.

Next, I added the login form to the website. I used HTML to create the form elements, including the username and password fields and the submit button.

I then used CSS to style the form to make it visually appealing and easy to use.